Stamina

+++ Assets/Scripts/Player/MagicBar.cs
@@ -7,12 +7,12 @@ public class MagicBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxMagic(int magic)
+    public void SetMaxMagic(float magic)
     {
         slider.maxValue = magic;
         slider.value = magic;
     }
-    public void SetMagic(int magic)
+    public void SetMagic(float magic)
     {
         slider.value = magic;
     }
